requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-binance>=1.0.19
ta>=0.11.0
gymnasium>=0.29.0
stable-baselines3>=2.2.0
torch>=2.1.0
joblib>=1.3.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
streamlit>=1.32.0
websocket-client>=1.7.0
urllib3<2
